Move github api into a separate module
