+++ all_repos/source/github.py
@@ -1,12 +1,7 @@
-import json
-import urllib.request
-from typing import Any
 from typing import Dict
-from typing import List
 from typing import NamedTuple
-from typing import Optional
-from typing import Tuple
-from typing import TypeVar
+
+from all_repos import github_api
 
 
 class Settings(NamedTuple):
@@ -17,45 +12,8 @@ class Settings(NamedTuple):
     private: bool = False
 
 
-def _parse_link_header(lnk: Optional[str]) -> Dict[str, str]:
-    if lnk is None:
-        return {}
-
-    ret = {}
-    parts = lnk.split(',')
-    for part in parts:
-        link, _, rel = part.partition(';')
-        link, rel = link.strip(), rel.strip()
-        assert link.startswith('<') and link.endswith('>'), link
-        assert rel.startswith('rel="') and rel.endswith('"'), rel
-        link, rel = link[1:-1], rel[len('rel="'):-1]
-        ret[rel] = link
-    return ret
-
-
-def _req(
-        url: str,
-        **kwargs: Any,
-) -> Tuple[List[Dict[str, Any]], Dict[str, str]]:
-    resp = urllib.request.urlopen(urllib.request.Request(url, **kwargs))
-    # TODO: mypy 620 should fix this
-    from typing import cast
-    resp = cast(urllib.response.addinfourl, resp)
-    return json.loads(resp.read()), _parse_link_header(resp.headers['link'])
-
-
-def _get_all(url: str, **kwargs: Any) -> List[Dict[str, Any]]:
-    ret: List[Dict[str, Any]] = []
-    resp, links = _req(url, **kwargs)
-    ret.extend(resp)
-    while 'next' in links:
-        resp, links = _req(links['next'], **kwargs)
-        ret.extend(resp)
-    return ret
-
-
 def list_repos(settings: Settings) -> Dict[str, str]:
-    repos = _get_all(
+    repos = github_api.get_all(
         'https://api.github.com/user/repos?per_page=100',
         headers={'Authorization': f'token {settings.api_key}'},
     )
@@ -68,17 +26,3 @@ def list_repos(settings: Settings) -> Dict[str, str]:
             (settings.collaborator or repo['permissions']['admin'])
         )
     }
-
-
-T = TypeVar('T', List[Any], Dict[str, Any], Any)
-
-
-def better_repr(obj: T) -> T:
-    if isinstance(obj, list):
-        return [better_repr(o) for o in obj]
-    elif isinstance(obj, dict):
-        return {
-            k: better_repr(v) for k, v in obj.items() if not k.endswith('url')
-        }
-    else:
-        return obj
